+++ src/main/java/com/invadermonky/pickuplimit/PickupLimits.java
@@ -17,7 +17,7 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 public class PickupLimits {
     public static final String MOD_ID = "pickuplimits";
     public static final String MOD_NAME = "Pickup Limits";
-    public static final String VERSION = "1.1.0";
+    public static final String VERSION = "1.2.0";
     public static final String MC_VERSION = "[1.12.2]";
 
     @Mod.Instance(MOD_ID)
@@ -25,7 +25,8 @@ public class PickupLimits {
 
     @Mod.EventHandler
     public void preInit(FMLPreInitializationEvent event) {
-        if (ModIds.crafttweaker.isLoaded()) MinecraftForge.EVENT_BUS.register(new ModIntegrationCT());
+        if (ModIds.crafttweaker.isLoaded())
+            MinecraftForge.EVENT_BUS.register(new ModIntegrationCT());
     }
 
     @Mod.EventHandler