+++ src/main/java/com/invadermonky/pickuplimit/limits/builders/AbstractLimitBuilder.java
@@ -6,8 +6,10 @@ import com.invadermonky.pickuplimit.limits.groups.AbstractLimitGroup;
 import com.invadermonky.pickuplimit.util.LogHelper;
 import com.invadermonky.pickuplimit.util.libs.ModIds;
 import gnu.trove.map.hash.THashMap;
+import it.unimi.dsi.fastutil.objects.Object2IntOpenHashMap;
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.potion.Potion;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.Tuple;
@@ -15,25 +17,24 @@ import net.minecraftforge.fml.common.Optional;
 
 import javax.annotation.Nullable;
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
 public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>, S extends AbstractLimitGroup<?>> {
     protected final String groupName;
     protected final int defaultLimit;
-    protected final NonNullList<ItemStack> groupStacks = NonNullList.create();
+    protected final NonNullList<Ingredient> groupIngredients = NonNullList.create();
+    protected final Map<ItemStack, Integer> armorLimitAdjustments = new Object2IntOpenHashMap<>();
+    protected final Map<ItemStack, Integer> baubleLimitAdjustments = new Object2IntOpenHashMap<>();
+    protected final Map<Tuple<Enchantment, Integer>, Integer> enchantmentLimitAdjustments = new Object2IntOpenHashMap<>();
+    protected final Map<Tuple<Potion, Integer>, Integer> potionLimitAdjustments = new Object2IntOpenHashMap<>();
+    protected final LinkedHashMap<String, Integer> stageLimitOverride = new LinkedHashMap<>();
+    protected final Map<String, NonNullList<Ingredient>> stagedStackRemovals = new THashMap<>();
+    protected final Map<Potion, Integer> encumberedEffects = new Object2IntOpenHashMap<>();
     protected ILimitFunction itemLimitFunction;
     protected String limitMessage;
     protected String limitTooltip;
     protected boolean allowOverLimit;
-    protected final Map<ItemStack, Integer> armorLimitAdjustments = new HashMap<>();
-    protected final Map<ItemStack, Integer> baubleLimitAdjustments = new HashMap<>();
-    protected final Map<Tuple<Enchantment, Integer>, Integer> enchantmentLimitAdjustments = new THashMap<>();
-    protected final Map<Tuple<Potion, Integer>, Integer> potionLimitAdjustments = new THashMap<>();
-    protected final LinkedHashMap<String, Integer> stageLimitOverride = new LinkedHashMap<>();
-    protected final THashMap<String, NonNullList<ItemStack>> stagedStackRemovals = new THashMap<>();
-    protected final THashMap<Potion, Integer> encumberedEffects = new THashMap<>();
 
     public AbstractLimitBuilder(String groupName, int defaultLimit) {
         this.groupName = groupName;
@@ -58,21 +59,21 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
     }
 
     /**
-     * Adds all passed ItemStacks to the group ItemStack list.
+     * Adds all passed Ingredients to the group ItemStack list.
      *
-     * @param stacks ItemStacks to add
+     * @param ingredients Ingredients to add
      * @return this
      */
-    public T addStacksToGroup(ItemStack... stacks) {
-        this.groupStacks.addAll(Arrays.asList(stacks));
+    public T addItemsToGroup(Ingredient... ingredients) {
+        this.groupIngredients.addAll(Arrays.asList(ingredients));
         return this.getThis();
     }
 
     /**
      * Returns all ItemStacks that have a limited pickup defined by this pickup limit
      */
-    public NonNullList<ItemStack> getGroupStacks() {
-        return this.groupStacks;
+    public NonNullList<Ingredient> getGroupItems() {
+        return this.groupIngredients;
     }
 
     public T setItemLimitValueFunction(ILimitFunction function) {
@@ -83,6 +84,16 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
     @Nullable
     public ILimitFunction getItemLimitFunction() {
         return this.itemLimitFunction;
+    }
+
+    /**
+     * Returns the defined pickup limit message string that will be displayed in the pickup fails. This can be a localization
+     * key or plain text.
+     *
+     * @return The pickup limit message string that will be displayed if the pickup fails.
+     */
+    public String getLimitMessage() {
+        return this.limitMessage;
     }
 
     /**
@@ -99,14 +110,8 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
         return this.getThis();
     }
 
-    /**
-     * Returns the defined pickup limit message string that will be displayed in the pickup fails. This can be a localization
-     * key or plain text.
-     *
-     * @return The pickup limit message string that will be displayed if the pickup fails.
-     */
-    public String getLimitMessage() {
-        return this.limitMessage;
+    public String getLimitTooltip() {
+        return this.limitTooltip;
     }
 
     public T setLimitTooltip(String limitTooltip) {
@@ -114,10 +119,6 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
             this.limitTooltip = limitTooltip;
         }
         return this.getThis();
-    }
-
-    public String getLimitTooltip() {
-        return this.limitTooltip;
     }
 
     public T setAllowOverlimit() {
@@ -134,7 +135,7 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
         return this.getThis();
     }
 
-    public THashMap<Potion, Integer> getEncumberedEffects() {
+    public Map<Potion, Integer> getEncumberedEffects() {
         return this.encumberedEffects;
     }
 
@@ -144,7 +145,7 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
      *
      * @param stack      The armor ItemStack
      * @param adjustment The limit adjustment when the armor is equipped
-     * @return
+     * @return this
      */
     public T addArmorLimitAdjustment(ItemStack stack, int adjustment) {
         if (!stack.isEmpty()) {
@@ -232,18 +233,18 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
     /**
      * Specifies any ItemStacks that will be removed from this group once the player obtains the specified GameStage.
      *
-     * @param stageName The name of the GameStage
-     * @param stacks    Any stacks that will be removed from this group limit when the player obtains the specified GameStage
+     * @param stageName   The name of the GameStage
+     * @param ingredients Any ingredients that will be removed from this group limit when the player obtains the specified GameStage
      * @return this
      */
     @Optional.Method(modid = ModIds.ConstIds.gamestages)
-    public T addStagedStackGroupRemoval(String stageName, ItemStack... stacks) {
-        NonNullList<ItemStack> stackList = NonNullList.from(ItemStack.EMPTY, stacks);
-        stackList.removeIf(ItemStack::isEmpty);
+    public T addStagedIngredientRemoval(String stageName, Ingredient... ingredients) {
+        NonNullList<Ingredient> ingredientList = NonNullList.from(Ingredient.EMPTY, ingredients);
+        ingredientList.removeIf(ingredient -> ingredient == Ingredient.EMPTY);
         if (this.stagedStackRemovals.containsKey(stageName)) {
-            this.stagedStackRemovals.get(stageName).addAll(stackList);
+            this.stagedStackRemovals.get(stageName).addAll(ingredientList);
         } else {
-            this.stagedStackRemovals.put(stageName, stackList);
+            this.stagedStackRemovals.put(stageName, ingredientList);
         }
         return this.getThis();
     }
@@ -252,7 +253,7 @@ public abstract class AbstractLimitBuilder<T extends AbstractLimitBuilder<T, S>,
      * Returns any defined GameStage group ItemStack removals.
      */
     @Optional.Method(modid = ModIds.ConstIds.gamestages)
-    public THashMap<String, NonNullList<ItemStack>> getStagedStackRemovals() {
+    public Map<String, NonNullList<Ingredient>> getStagedStackRemovals() {
         return this.stagedStackRemovals;
     }