+++ src/main/java/com/invadermonky/pickuplimit/limits/groups/PickupLimitGroup.java
@@ -6,8 +6,8 @@ import com.invadermonky.pickuplimit.limits.builders.PickupLimitBuilder;
 import com.invadermonky.pickuplimit.limits.caches.AbstractGroupCache;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.potion.PotionEffect;
-import net.minecraftforge.oredict.OreDictionary;
 
 public class PickupLimitGroup extends AbstractLimitGroup<PickupLimitBuilder> {
     public static final PickupLimitGroup EMPTY = new PickupLimitBuilder(null, -1).build();
@@ -25,11 +25,9 @@ public class PickupLimitGroup extends AbstractLimitGroup<PickupLimitBuilder> {
     public boolean matches(EntityPlayer player, ItemStack stack) {
         //Values of -1 means the limit is not active.
         if (!stack.isEmpty() && this.getLimitWithStageOverride(player) != -1) {
-            for (ItemStack limitStack : this.getLimitStacks(player)) {
-                if (OreDictionary.itemMatches(limitStack, stack, false)) {
-                    if (!limitStack.hasTagCompound() || ItemStack.areItemStackTagsEqual(stack, limitStack)) {
-                        return true;
-                    }
+            for (Ingredient limitIngredient : this.getLimitIngredients(player)) {
+                if (limitIngredient.apply(stack)) {
+                    return true;
                 }
             }
         }