+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 <br><br>
+## [1.12.2-1.2.0]
+### Changed
+- Renamed `addStacks()` builder method to `addItems()`. This method now supports IIngredients.
+- Renamed Pickup Limit Builder `addStagedStackRemovals()` to `addStagedItemRemovals()`. This method now supports IIngredients.
+### Fixed
+- Fixed over-limit items being picked up and dropped constantly
+### Removed
+- Removed Pickup Limit Builder `addOreDict()` method. This is functionality is included in `addItems()`.
+- Removed Pickup Limit Builder `addStagedOreRemovals()` method. This functionality is included in `addStagedItemRemovals()`.
+
+---
 
 ## [1.12.2-1.1.0]
 ### Added
@@ -25,8 +36,8 @@
 - fixed a circular logic error when using `ILimitGroup#getStackLimitValue(ItemStack)`
   - `ILimitGroup#getStackLimitValue(ItemStack)` has been renamed to `ILimitGroup#getDefaultLimitValue(ItemStack)`
 - fixed an Item Lifetime parsing error when setting values to -32768
-<br><br>
 
+---
 
 ## [1.12.2-1.0.0]
 - Initial Release